requirements.txt
agno>=2.3.26
groq>=1.0.0
openai>=2.15.0
python-dotenv>=1.2.1
tavily-python>=0.7.17
yfinance>=1.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
